Remove unnecessary calls to float() before division.
Most of these calls probably date back to the time before
`__future__`-division existed...

Note that this will have the side effect that some functions that were
accidentally accepting strings instead of floats before will no longer
accept strings -- which seems fine to me.

+++ examples/statistics/hist.py
@@ -52,7 +52,7 @@ fig, axs = plt.subplots(1, 2, tight_layout=True)
 N, bins, patches = axs[0].hist(x, bins=n_bins)
 
 # We'll color code by height, but you could use any scalar
-fracs = N.astype(float) / N.max()
+fracs = N / N.max()
 
 # we need to normalize the data to 0..1 for the full range of the colormap
 norm = colors.Normalize(fracs.min(), fracs.max())
